% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IhateNYTSpellBee-package.R
\docType{package}
\name{IhateNYTSpellBee-package}
\alias{IhateNYTSpellBee}
\alias{IhateNYTSpellBee-package}
\title{IhateNYTSpellBee: Generate answers for NYT Spelling Bee}
\description{
Package to kill the enjoyment of the 
      NYT spelling bee.
}
\keyword{NYT}
