% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KillBee.R
\name{kill_the_bee}
\alias{kill_the_bee}
\title{Kill the joy of NYT spelling bee}
\usage{
kill_the_bee(key, circle)
}
\arguments{
\item{key}{Middle letter}

\item{circle}{Other letters around circle as a string}
}
\value{
vector of words from scrabble dictionary that fit requirements
}
\description{
Removes the enjoyment of the NYT spelling bee app by giving you the answers
}
\examples{
library(IhateNYTSpellBee)
  kill_the_bee("f", "ailmce")
}
